Added functionality to configure settings for spawning nets
The system now allows you to define specific settings for each mesh that are applied when it is spawned. This makes it possible to fine-tune the behavior of meshes, such as the inclusion of function blocks, attenuator settings, editability, and static and power status, according to the needs of the event. This change allows for more flexibility and control over the generated meshes.